Create Exercise 1: User Input Output
Write code to prompt the user to input her/his name and age and print these details on the screen. Find the length of the name and also the age of the user after one year. The format of text should look like the sample output below. (Use title() function)