Create Exercise C: Calculator Function
The program should display the following calculator menu:

1. Add
2. Subtract
3. Multiply
4. Divide
5. Modulus
The program will prompt the user to choose the operation choice (from 1 to 5). Then it asks the user to input values for the calculation. The program outputs the results of the calculation. The program should end by asking if the user would like to perform another calculation or not.